requirements.txt
Flask==3.0.3
mysql-connector-python==8.0.33
